requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
djangorestframework==3.14.0